+++ example/android/app/src/main/kotlin/dev/joshi/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package dev.joshi.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}