+++ lib/src/chess_board_controller.dart
@@ -1,89 +1,150 @@
-import 'package:chess/chess.dart' as chess;
+import 'package:chess/chess.dart';
+import 'package:flutter/material.dart';
+import 'constants.dart';
 
-enum PieceType { Pawn, Rook, Knight, Bishop, Queen, King }
+class ChessBoardController extends ValueNotifier<Chess> {
+  late Chess game;
 
-enum PieceColor {
-  White,
-  Black,
-}
+  factory ChessBoardController() => ChessBoardController._(Chess());
 
-/// Controller for programmatically controlling the board
-class ChessBoardController {
-  /// The game attached to the controller
-  chess.Chess game;
+  factory ChessBoardController.fromGame(Chess game) =>
+      ChessBoardController._(game);
 
-  /// Function from the ScopedModel to refresh board
-  Function refreshBoard;
+  factory ChessBoardController.fromFEN(String fen) =>
+      ChessBoardController._(Chess.fromFEN(fen));
 
-  /// Makes move on the board
-  void makeMove(String from, String to) {
-    game?.move({"from": from, "to": to});
-    refreshBoard == null ? this._throwNotAttachedException() : refreshBoard();
-  }
+  ChessBoardController._(Chess game)
+      : game = game,
+        super(game);
 
   /// Makes move on the board
-  void makeMoveWithNormalNotation(String move) {
-    game?.move(move);
-    refreshBoard == null ? this._throwNotAttachedException() : refreshBoard();
+  void makeMove({required String from, required String to}) {
+    game.move({"from": from, "to": to});
+    notifyListeners();
   }
 
   /// Makes move and promotes pawn to piece (from is a square like d4, to is also a square like e3, pieceToPromoteTo is a String like "Q".
   /// pieceToPromoteTo String will be changed to enum in a future update and this method will be deprecated in the future
-  void makeMoveWithPromotion(String from, String to, String pieceToPromoteTo) {
-    game?.move({"from": from, "to": to, "promotion": pieceToPromoteTo});
-    refreshBoard == null ? this._throwNotAttachedException() : refreshBoard();
+  void makeMoveWithPromotion(
+      {required String from,
+      required String to,
+      required String pieceToPromoteTo}) {
+    game.move({"from": from, "to": to, "promotion": pieceToPromoteTo});
+    notifyListeners();
   }
 
-  /// Resets square
+  /// Makes move on the board
+  void makeMoveWithNormalNotation(String move) {
+    game.move(move);
+    notifyListeners();
+  }
+
+  void undoMove() {
+    if (game.half_moves == 0) {
+      return;
+    }
+    game.undo_move();
+    notifyListeners();
+  }
+
   void resetBoard() {
-    game?.reset();
-    refreshBoard == null ? this._throwNotAttachedException() : refreshBoard();
+    game.reset();
+    notifyListeners();
   }
 
   /// Clears board
   void clearBoard() {
-    game?.clear();
-    refreshBoard == null ? this._throwNotAttachedException() : refreshBoard();
+    game.clear();
+    notifyListeners();
   }
 
   /// Puts piece on a square
-  void putPiece(PieceType piece, String square, PieceColor color) {
-    game?.put(_getPiece(piece, color), square);
-    refreshBoard == null ? this._throwNotAttachedException() : refreshBoard();
+  void putPiece(BoardPieceType piece, String square, PlayerColor color) {
+    game.put(_getPiece(piece, color), square);
+    notifyListeners();
   }
 
   /// Loads a PGN
   void loadPGN(String pgn) {
     game.load_pgn(pgn);
-    refreshBoard == null ? this._throwNotAttachedException() : refreshBoard();
+    notifyListeners();
   }
 
-  /// Exception when a controller is not attached to a board
-  void _throwNotAttachedException() {
-    throw Exception("Controller not attached to a ChessBoard widget!");
+  /// Loads a PGN
+  void loadFen(String fen) {
+    game.load(fen);
+    notifyListeners();
+  }
+
+  bool isCheckMate() {
+    return game.in_checkmate;
+  }
+
+  bool isDraw() {
+    return game.in_draw;
+  }
+
+  bool isStaleMate() {
+    return game.in_stalemate;
+  }
+
+  bool isThreefoldRepetition() {
+    return game.in_threefold_repetition;
+  }
+
+  bool isInsufficientMaterial() {
+    return game.insufficient_material;
+  }
+
+  bool isGameOver() {
+    return game.game_over;
+  }
+
+  String getAscii() {
+    return game.ascii;
+  }
+
+  String getFen() {
+    return game.fen;
+  }
+
+  List<String?> getSan() {
+    return game.san_moves();
+  }
+
+  List<Piece?> getBoard() {
+    return game.board;
+  }
+
+  List<Move> getPossibleMoves() {
+    return game.moves({'asObjects': true}) as List<Move>;
+  }
+
+  int getMoveCount() {
+    return game.move_number;
+  }
+
+  int getHalfMoveCount() {
+    return game.half_moves;
   }
 
   /// Gets respective piece
-  chess.Piece _getPiece(PieceType piece, PieceColor color) {
-    chess.Color _getColor(PieceColor color) {
-      return color == PieceColor.White ? chess.Color.WHITE : chess.Color.BLACK;
-    }
+  Piece _getPiece(BoardPieceType piece, PlayerColor color) {
+    var convertedColor = color == PlayerColor.white ? Color.WHITE : Color.BLACK;
 
     switch (piece) {
-      case PieceType.Bishop:
-        return chess.Piece(chess.PieceType.BISHOP, _getColor(color));
-      case PieceType.Queen:
-        return chess.Piece(chess.PieceType.QUEEN, _getColor(color));
-      case PieceType.King:
-        return chess.Piece(chess.PieceType.KING, _getColor(color));
-      case PieceType.Knight:
-        return chess.Piece(chess.PieceType.KNIGHT, _getColor(color));
-      case PieceType.Pawn:
-        return chess.Piece(chess.PieceType.PAWN, _getColor(color));
-      case PieceType.Rook:
-        return chess.Piece(chess.PieceType.ROOK, _getColor(color));
+      case BoardPieceType.Bishop:
+        return Piece(PieceType.BISHOP, convertedColor);
+      case BoardPieceType.Queen:
+        return Piece(PieceType.QUEEN, convertedColor);
+      case BoardPieceType.King:
+        return Piece(PieceType.KING, convertedColor);
+      case BoardPieceType.Knight:
+        return Piece(PieceType.KNIGHT, convertedColor);
+      case BoardPieceType.Pawn:
+        return Piece(PieceType.PAWN, convertedColor);
+      case BoardPieceType.Rook:
+        return Piece(PieceType.ROOK, convertedColor);
     }
-
-    return chess.Piece(chess.PieceType.PAWN, chess.Color.WHITE);
   }
 }